% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseflow_data_used_in_first_round_of_SFR.rds}
\alias{baseflow_data_used_in_first_round_of_SFR.rds}
\title{[DATASET] Baseflow used in initial MODFLOW/SFR application}
\format{A data frame with 874453 rows and 4 variables:
\itemize{
  \item{\code{siteNo}}{character USGS gage ID.}
  \item{\code{Date}}{date Date associated with baseflow estimate. }
  \item{\code{baseFlow}}{double Estimated baseflow, in cubic feet per second.}
  \item{\code{comment}}{character One of 'estimated gaged', 'estimated ungaged', or 'calculated'.}
}}
\usage{
readr::read_rds("data/baseflow_data_used_in_first_round_of_SFR.rds")
}
\description{
Original random forest model output as supplied to the MODFLOW modelers.
}
\details{
This is the data file used in the initial application of the SFR MODFLOW package as applied to the
MERAS study area. The file contains a mix of random forest model forecasts and observed values for baseflow.
Values with a comment field value of 'estimated ungaged' or 'estimated gaged' represent outputs from the random
forest model. Values with the comment field listed as 'calculated' represent observed values of baseflow
as calculated from streamflow records using hydrograph separation methods.
}
\keyword{datasets}
