% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dvsMonthly.rds}
\alias{dvsMonthly.rds}
\title{[DATASET] ORIGINAL monthly streamflow and baseflow summaries}
\format{A file containing 180989 observations of 5 variables}
\usage{
readr::read_rds("data/dvsMonthly.rds")
}
\description{
This dataset contains monthly streamflow and baseflow summaries for a list of USGS gage IDs pulled
from a shapefile of gage drainage basins. The dataset contains almost 21,300 missing or negative
 flow values, which were removed in subsequent processing in the original scripts.
}
\details{
This dataset also includes results for additional gage IDs that were removed in the original scripts
before training the random forest trees. The sites removed in subsequent processing were:

"03320500","03383000","03434500","03436100","03438000","03438220",
  "03588000","03601990","03602500","03603000","03604400","03605555",
  "03436690","03436700","03584000","03584020","03588400","03588500",
  "03599450","03600500","03601630","03604000","03433641"

The variables are as follows:

\itemize{
  \item siteNo. USGS gage ID
  \item Date. Year and month associated with summary statistic (YYYY-MM)
  \item Flow. Monthly mean streamflow in cubic feet per second
  \item baseFlow. Stream baseflow as estimated by the 'part' method of baseflow separation, in cubic feet per second
  \item isSubbed. Internal variable used by the original scripts to note whether a value is observed or imputed ('real' or 'fake')
}
}
\keyword{datasets}
