% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlowDataForSite.R
\name{getFlowDataForSite}
\alias{getFlowDataForSite}
\title{Retrieve and summarize flow data by month for a site}
\usage{
getFlowDataForSite(site)
}
\arguments{
\item{site}{A character representation of a USGS gage ID}
}
\value{
List containing 1) 'df': a dataframe listing the mean streamflow and baseflow where calculable for each month-year in the record, and 2) 'siteinfo': site information returned from the USGS NWIS system
}
\description{
This function retrieves site data for a USGS gaging station, pulls down all daily mean values
that exist in the record, and summarizes the mean daily flow by month. In addition, the function
calls the 'part' routine coded as part of the USGS DVstats package to perform a baseflow separation
on the daily values. Months for which all or part of the daily values are 'NA' (missing) are deleted.
}
\examples{
mylist <- getFlowDataForSite("07063000")
head(mylist$df)
print(mylist$siteinfo)
}
