% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRasterMean.R
\name{getRasterMean}
\alias{getRasterMean}
\title{Calculate mean raster value for a polygon feature}
\usage{
getRasterMean(rasterFilename, shapeFile)
}
\arguments{
\item{rasterFilename}{A string identifying the complete path and filename of a GDAL-readable raster}

\item{shapeFile}{A shapefile object of class 'sp' containing the polygons for which zone calculations will be made}
}
\value{
rasterMean A scalar or array of raster mean values associated with the polygon areas
}
\description{
Returns the mean value of raster pixels that fall within a
simple shapefile boundary.
}
\examples{
# read in the precipitation raster and the basin outline shapefile
myRasterFile <- system.file("extdata", "PRISM_ppt_stable_4kmM2_190004_bil.bil", package="mapRandomForest")
myShapeFile <- system.file("extdata", "ungagedSheds.shp", package = "mapRandomForest")
myRaster <- raster::raster(myRasterFile)
myShapes <- raster::shapefile(myShapeFile)

# reproject the shapefile to lat/lon and extract centroid values
myShapes_latlon <- sp::spTransform(myShapes, myRaster@crs)

# calculate the mean raster values associated with each polygon in shapefile
precip_mm <- getRasterMean(myRaster, myShapes_latlon)

# tack the mean precipitation values onto the data table of the shapefile
myShapes@data <- cbind(myShapes@data,precip_mm)
}
