% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hargreaves_samani_ET0.R
\name{hargreavesET0}
\alias{hargreavesET0}
\title{Calculate reference ET0 by means of the Hargreaves-Samani relation}
\usage{
hargreavesET0(tmin, tmax, day_of_year, days_in_month, precipitation, lat)
}
\arguments{
\item{tmin}{Minimum air temperature in degrees C}

\item{tmax}{Maximum air temperature in degrees C}

\item{day_of_year}{Julian Day (day-of-year, January 1=1)}

\item{days_in_month}{Number of days in month associated with Julian Day}

\item{precipitation}{Precipitation in millimeters}

\item{lat}{Latitude in decimal degrees}
}
\value{
ET0 The reference ET0 value in millimeters per month
}
\description{
Calculate reference ET0 by means of the Hargreaves-Samani relation
}
