% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyContiguous.R
\name{identify_contiguous}
\alias{identify_contiguous}
\title{Identify contiguous blocks of data}
\usage{
identify_contiguous(date_strings)
}
\arguments{
\item{yyyymm}{A string representing the date expressed as "YYYY-MM"}
}
\value{
blockIDs A vector of numerical block IDs
}
\description{
Returns a vector identifying contigous blocks of data. Data is judged
to be part of a contiguous block if the date (expressed as year and month)
increases monotonically without gaps. The block ID is incremented by one
when a gap is identified.
}
