% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPRISMRasters.R
\name{processPRISMRasters}
\alias{processPRISMRasters}
\title{Calculate monthly mean raster values for a set of polygons}
\usage{
processPRISMRasters(shp, rasterDir, dateSeq)
}
\arguments{
\item{shp}{An sp object created from a polygon shapefile}

\item{rasterDir}{Fully qualified directory name containing PRISM *.bil files for processing}

\item{dateSeq}{A list of POSIX date objects defining the YYYY-MM}
}
\description{
Calculate monthly mean raster values for a set of polygons
}
