% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sitesAll.shp}
\alias{sitesAll.shp}
\title{[DATASET] USGS gage contributing areas}
\format{A shapefile and data frame with 316 rows and 3 variables.
\describe{
  \item{\code{site_no}}{character USGS gaging station ID}
  \item{\code{Area}}{double Contributing area in square miles}
}

Attributes of this shapefile include:

\describe{
  \item{\code{class}}{SpatialPolygonsDataFrame}
  \item{\code{features}}{316}
  \item{\code{extent}}{-36585, 970995, 791895, 1741095  (xmin, xmax, ymin, ymax)}
  \item{\code{crs}}{+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0}
  \item{\code{variables}}{2}
  \item{\code{names}}{site_no, Area}
  \item{\code{min values}}{02374250, 2.03630278442}
  \item{\code{max values}}{08041500, 18426.5671397}
}}
\source{
\url{https://water.usgs.gov/lookup/getspatial?gagesII_Sept2011}
}
\usage{
raster::shapefile("data/sitesAll.shp")
}
\description{
Shapefile of contributing areas associated with USGS gages
}
\examples{
myShapeFile <- system.file("data", "sitesAll.shp", package = "mapRandomForest")

myBasins <- raster::shapefile(myShapeFile)
plot(myBasins)
}
