% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testFlowDat.rds}
\alias{testFlowDat.rds}
\title{[DATASET] Unedited ORIGINAL random forest model inputs}
\format{A file containing 158526 observations of 39 variables}
\usage{
readr::read_rds("data/testFlowDat.rds")
}
\description{
This dataset was used as raw input to the random forest model scripts. The dataset combines the output
from several scripts, joining data flow and baseflow summaries with precipitation, air temperature,
location, elevation, and generalized geologic units.
}
\details{
In later processing steps, the fields pertaining to the digital elevation map statistics and to
basin geology were dropped from the list of random forest model inputs.

\itemize{
\item siteNo. USGS gage ID
\item Date. Date of discharge, precipitation, and temperature observation
\item Flow. Total daily average discharge, in cubic feet per second
\item baseFlow. Estimated daily average baseflow calculated by means of the HYSEP algorithm, cubic feet per second
\item precip. Daily average zonal precipitation calculated for the drainage basin associated with the gage, in millimeters
\item tmax. Daily average zonal maximum air temperature calculated for the drainage basin associated with the gage, in degrees Celcius
\item tmean. Daily average zonal mean air temperature calculated for the drainage basin associated with the gage, in degrees Celcius
\item tmin. Daily average zonal minimum air temperature calculated for the drainage basin associated with the gage, in degrees Celcius
\item X. Longitude of gaging basin centroid, decimal degrees
\item Y. Latitude of gaging basin centroid, decimal degrees
\item ET0Har. Reference evapotranspiration calculated by means of the Hargreaves-Samani relation, in millimeters
\item preSub1. Value of `precip` from previous month, in millimeters
\item preSub2. Value of `precip` from two months prior to current month, in millimeters
\item tmaxSub1. Value of `tmax` from previous month, in degrees Celcius
\item tmaxSub2. Value of `tmax` from two months prior to current month, in degrees Celcius
\item tmeanSub1. Value of `tmean` from previous month, in degrees Celcius
\item tmeanSub2. Value of `tmean` from two months prior to current month, in degrees Celcius
\item tminSub1. Value of `tmin` from previous month, in degrees Celcius
\item tminSub2. Value of `tmin` from two months prior to current month, in degrees Celcius
\item ET0Sub1. Value of `ET0Har` from previous month, in millimeters
\item ET0Sub2. Value of `ET0Har` from two months prior to current month, in millimeters
\item preTot6. Value of `precip` summed over previous 6 months, in millimeters
\item MIN. Minimum elevation of the gaging basin, in meters
\item MAX. Maximum elevation of the gaging basin, in meters
\item RANGE. Range of elevation of the gaging basin, in meters
\item MEAN. Mean elevation of the gaging basin, in meters
\item STD. Standard deviation of the elevation of the gaging basin, in meters
\item H2o. Percentage of gaging basin classified as open water
\item Q. Percentage of gaging basin having geology classified as Quaternary
\item pgT. Percentage of gaging basin having geology classified as Paleogene
\item K. Percentage of gaging basin having geology classified as Cretaceous
\item lPz. Percentage of gaging basin having geology classified as Lower Paleozoic
\item nT. Percentage of gaging basin having geology classified as Neogene
\item Yv. Percentage of gaging basin having geology classified as Middle Proterozoic
\item mPz. Percentage of gaging basin having geology classified as Middle Paleozoic
\item uPz. Percentage of gaging basin having geology classified as Upper Paleozoic
\item Yg. Percentage of gaging basin having geology classified as Middle Proterozoic
\item ZPz. Percentage of gaging basin having geology classified as Late Proterozoic
\item drain_area_va. Drainage area of the drainage basin associated with the stream gage, in square miles
}
}
\keyword{datasets}
